いいね機能のテストをfavorites_spec.rbから各システムスペックに移行
