+++ spec/system/homes_spec.rb
@@ -85,7 +85,7 @@ RSpec.describe 'Homes', type: :system do
     end
   end
 
-  describe 'ランキング表示とリンクのテスト' do
+  describe 'ランキングのテスト' do
     let(:user) { create(:user, name: 'shumpei') }
     let!(:first_store) {
       create(:store, :rated1, name: 'あかちゃん本舗', introduction: '綺麗な授乳室でした', postcode: '1111111', prefecture_code: '北海道',
@@ -100,75 +100,116 @@ RSpec.describe 'Homes', type: :system do
                               prefecture_code: '北海道', city: '稚内市1-1-1', user: user)
     }
 
-    before do
-      visit root_path
+    describe '表示とリンクのテスト' do
+      before do
+        visit root_path
+      end
+
+      it '全ての施設情報がレビュー平均点順に表示されていること' do
+        within '.store-1' do
+          expect(page).to have_selector '.card-rank', text: '第1位'
+          expect(page).to have_selector('img[alt=施設画像-1]')
+          expect(page).to have_selector '.card-title', text: 'あかちゃん本舗'
+          expect(page).to have_css ".favorite-#{first_store.id}"
+          expect(page).to have_selector '.favorite-count', text: '0'
+          expect(page).to have_css '.review-average-rating'
+          expect(page).to have_selector '.reviews-average-score', text: '4.0'
+          expect(page).to have_selector '.reviews-count', text: '3'
+          expect(page).to have_content '綺麗な授乳室でした'
+          expect(page).to have_content '北海道函館市1-1-1'
+          expect(page).to have_link 'shumpei'
+        end
+
+        within '.store-2' do
+          expect(page).to have_selector '.card-rank', text: '第2位'
+          expect(page).to have_selector('img[alt=施設画像-2]')
+          expect(page).to have_selector '.card-title', text: 'ベビーレストラン'
+          expect(page).to have_css ".favorite-#{second_store.id}"
+          expect(page).to have_selector '.favorite-count', text: '0'
+          expect(page).to have_css '.review-average-rating'
+          expect(page).to have_selector '.reviews-average-score', text: '3.0'
+          expect(page).to have_selector '.reviews-count', text: '3'
+          expect(page).to have_content '個室の和室があって赤ちゃんと一緒でもゆっくりできました'
+          expect(page).to have_content '沖縄県那覇市1-1-1'
+          expect(page).to have_link 'shumpei'
+        end
+
+        within '.store-3' do
+          expect(page).to have_selector '.card-rank', text: '第3位'
+          expect(page).to have_selector('img[alt=施設画像-3]')
+          expect(page).to have_selector '.card-title', text: '東松屋'
+          expect(page).to have_css ".favorite-#{third_store.id}"
+          expect(page).to have_selector '.favorite-count', text: '0'
+          expect(page).to have_css '.review-average-rating'
+          expect(page).to have_selector '.reviews-average-score', text: '2.0'
+          expect(page).to have_selector '.reviews-count', text: '3'
+          expect(page).to have_content '広くて新しい授乳室でした'
+          expect(page).to have_content '北海道稚内市1-1-1'
+          expect(page).to have_link 'shumpei'
+        end
+        expect(page).to have_selector('.store-card', count: 3)
+      end
+
+      it '施設画像をクリックすると施設詳細画面にページ遷移すること' do
+        click_link '施設画像-1'
+        expect(current_path).to eq store_path(first_store)
+        expect(page).to have_selector 'h1', text: 'あかちゃん本舗'
+      end
+
+      it 'コメントアイコンをクリックすると施設詳細画面にページ遷移すること' do
+        within '.store-1' do
+          find('.comment-link').click
+        end
+        expect(current_path).to eq store_path(first_store)
+        expect(page).to have_selector 'h1', text: 'あかちゃん本舗'
+      end
+
+      it 'ユーザー名をクリックするとユーザーの詳細画面に遷移すること' do
+        within '.store-1' do
+          click_link 'shumpei'
+        end
+        expect(current_path).to eq user_path(user)
+        expect(page).to have_selector '.card-title', text: 'shumpei'
+      end
     end
 
-    it '全ての施設情報がレビュー平均点順に表示されていること' do
-      within '.store-1' do
-        expect(page).to have_selector '.card-rank', text: '第1位'
-        expect(page).to have_selector('img[alt=施設画像-1]')
-        expect(page).to have_selector '.card-title', text: 'あかちゃん本舗'
-        expect(page).to have_css ".favorite-#{first_store.id}"
-        expect(page).to have_selector '.favorite-count', text: '0'
-        expect(page).to have_css '.review-average-rating'
-        expect(page).to have_selector '.reviews-average-score', text: '4.0'
-        expect(page).to have_selector '.reviews-count', text: '3'
-        expect(page).to have_content '綺麗な授乳室でした'
-        expect(page).to have_content '北海道函館市1-1-1'
-        expect(page).to have_link 'shumpei'
+    describe 'いいね機能のテスト', js: true do
+      context 'ログインしているとき' do
+        it '.favorite-store.idをクリックするといいねをつけたり削除したりできること' do
+          sign_in(user)
+          page.accept_confirm do
+            visit root_path
+          end
+          within '.store-1' do
+            expect(page).to have_selector '.favorite-count', text: '0'
+
+            expect {
+              find(".favorite-#{first_store.id}").click
+              expect(page).to have_css ".favorited-#{first_store.id}"
+              expect(page).to have_selector '.favorite-count', text: '1'
+            }.to change(Favorite, :count).by(1)
+
+            expect {
+              find(".favorited-#{first_store.id}").click
+              expect(page).to have_css ".favorite-#{first_store.id}"
+              expect(page).to have_selector '.favorite-count', text: '0'
+            }.to change(Favorite, :count).by(-1)
+          end
+        end
       end
 
-      within '.store-2' do
-        expect(page).to have_selector '.card-rank', text: '第2位'
-        expect(page).to have_selector('img[alt=施設画像-2]')
-        expect(page).to have_selector '.card-title', text: 'ベビーレストラン'
-        expect(page).to have_css ".favorite-#{second_store.id}"
-        expect(page).to have_selector '.favorite-count', text: '0'
-        expect(page).to have_css '.review-average-rating'
-        expect(page).to have_selector '.reviews-average-score', text: '3.0'
-        expect(page).to have_selector '.reviews-count', text: '3'
-        expect(page).to have_content '個室の和室があって赤ちゃんと一緒でもゆっくりできました'
-        expect(page).to have_content '沖縄県那覇市1-1-1'
-        expect(page).to have_link 'shumpei'
+      context 'ログインしていないとき' do
+        it '.favorite-store.idをクリックするとログインページに遷移すること' do
+          page.accept_confirm do
+            visit root_path
+          end
+          within '.store-1' do
+            find(".favorite-#{first_store.id}").click
+          end
+          expect(current_path).to eq new_user_session_path
+          expect(page).to have_selector '.alert-alert', text: 'ログインもしくはアカウント登録してください'
+        end
       end
-
-      within '.store-3' do
-        expect(page).to have_selector '.card-rank', text: '第3位'
-        expect(page).to have_selector('img[alt=施設画像-3]')
-        expect(page).to have_selector '.card-title', text: '東松屋'
-        expect(page).to have_css ".favorite-#{third_store.id}"
-        expect(page).to have_selector '.favorite-count', text: '0'
-        expect(page).to have_css '.review-average-rating'
-        expect(page).to have_selector '.reviews-average-score', text: '2.0'
-        expect(page).to have_selector '.reviews-count', text: '3'
-        expect(page).to have_content '広くて新しい授乳室でした'
-        expect(page).to have_content '北海道稚内市1-1-1'
-        expect(page).to have_link 'shumpei'
-      end
-      expect(page).to have_selector('.store-card', count: 3)
-    end
-
-    it '施設画像をクリックすると施設詳細画面にページ遷移すること' do
-      click_link '施設画像-1'
-      expect(current_path).to eq store_path(first_store)
-      expect(page).to have_selector 'h1', text: 'あかちゃん本舗'
-    end
-
-    it 'コメントアイコンをクリックすると施設詳細画面にページ遷移すること' do
-      within '.store-1' do
-        find('.comment-link').click
-      end
-      expect(current_path).to eq store_path(first_store)
-      expect(page).to have_selector 'h1', text: 'あかちゃん本舗'
-    end
-
-    it 'ユーザー名をクリックするとユーザーの詳細画面に遷移すること' do
-      within '.store-1' do
-        click_link 'shumpei'
-      end
-      expect(current_path).to eq user_path(user)
-      expect(page).to have_selector '.card-title', text: 'shumpei'
     end
   end
 end
